requirements.txt
transformers==4.40.0
torch==2.3.0
fastapi==0.111.0
uvicorn==0.29.0
plotly==5.22.0
dash==2.17.0
dash-bootstrap-components==1.6.0
yfinance==0.2.40
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.31.0
pytest==7.4.0
pytest-mock==3.14.0
httpx==0.27.0
